frontend/requirements.txt
# Streamlit
streamlit==1.38.0

# HTTP client
httpx==0.27.2
requests==2.32.3

# Image handling
Pillow==10.4.0

# Utilities
python-dotenv==1.0.1

# Charts (optional)
plotly==5.24.1
